Added analytics (#26)
* Integrated hotjar properly

* Disable delete account for now

* This update includes changes to several files in the application. The changes involve removing the autoComplete attribute from various
input fields in the forms. The autoComplete attribute is set to an empty string (""), effectively disabling autocomplete for those fields.
This change ensures that the browser does not suggest previously entered values for those input fields. The affected files include
CollectionForm.tsx, FunctionArgumentFormField.tsx, FunctionForm.tsx, UserProfileForm.tsx, and MagicLoginForm.tsx. The purpose of this
update is to improve user experience and prevent unintended data leakage through autocomplete suggestions.

* This update introduces a change to the UserProfileForm.tsx file and the updateProfile function in the profile.ts file.

In UserProfileForm.tsx, a new variable lockUsername is introduced, which is set to true if the userProfile.userName exists. This variable
is then used to set the readOnly and disabled attributes of the Input component for the userName field.

In the updateProfile function in profile.ts, a check is added to prevent updating the userName if it already exists in the database. If
the existing userName is different from the updated userName, an error is thrown, indicating that the username cannot be updated once set
and suggesting contacting support for assistance.

* Mixpanel initial implementation for collections,functions and subscriptions api calls

* File locations refactor for vm function calls

* Updated lock file

* Mixpanel implementation for collections, functions and subscriptions api calls #2

* Mixpanel implementation for built-ins

* Refactor analytics and middleware code

This commit refactors the code in the analytics/index.ts file and the middleware.ts file.

In analytics/index.ts, the change adds a new constant IS_PROD from @lib/constants and modifies the mixpanel export. Now, mixpanel is
initialized with MIXPANEL_TOKEN only if IS_PROD is true. Otherwise, it is initialized with a dummy object that has a track method that
does nothing. This change allows for better handling of analytics in production and non-production environments.

In middleware.ts, the change removes the setting of the x-next-pathname header and updates the condition for logging the pathname. Now,
the logging is done if the pathname includes '/api/v1'. This change improves the middleware logic for API versioning.