Refactor code and remove unused imports and console.log statements
This commit refactors the code in the functions.ts file by removing the unused import of MIXPANEL_TOKEN from the @lib/constants module.
Additionally, it removes the console.log statement that logs the Mixpanel token. These changes aim to clean up the codebase and remove
unnecessary code.

In the utils.ts file, the commit removes the geo property from the analytics object returned by the getAnalyticsObject function. This
property was not being used and was therefore removed to simplify the code.