Fix openapi schema (#33)
* Update session cookie name in .env.example and .env.live

This commit updates the session cookie name in the .env.example and .env.live files. The previous session cookie name ING_RAAAAAA has been changed
to ING_SESSION. This change ensures consistency between the environment variables used in the code and the example and live environment files. The
APP_SESSION_COOKIE_NAME constant in constants.ts has also been updated accordingly.

* Refactor imports in ChatForm.tsx

This commit refactors the imports in the ChatForm.tsx file. It removes the unused imports "useAssistant" and "APP_URL, BAKA_ASSISTANT_ROOT_PATH,
USER_API_ROOT_PATH" from the "ai/react" and "@lib/constants" modules respectively. The "useAssistant" import is no longer used in the code, and the
unused constants are removed. This improves code readability and reduces unnecessary dependencies.

* Refactor user API URLs in the openapi/page.tsx file

This commit refactors the user API URLs in the openapi/page.tsx file. Previously, the URLs were pointing to the root of the user API, but now they
have been updated to include the "/me" path. This change ensures that the API requests are made to the correct endpoints for the authenticated
user. The userOpenApiUrl variable now points to the "/me/openapi.json" endpoint, and the userSwaggerUrl variable points to the "/me/swagger"
endpoint. This update improves the accuracy and security of the API calls in the application.

* This update includes several changes in the codebase.

 1 In the page.tsx file, the import statement for USER_API_ROOT_PATH is replaced with ME_API_ROOT_PATH from the constants.ts file.
 2 In the layout.tsx file, the div element with id overview-layout is given the h-full and w-full classes.
 3 In the OpenAPISpecLoader.tsx file, the cn and ScrollArea components are imported from the lib/utils and ui/scroll-area respectively. The
   createPathTree function is modified to use ME_API_ROOT_PATH instead of USER_API_ROOT_PATH. Some CSS classes are added to the div elements.
 4 In the page.tsx file, the import statements for USER_API_ROOT_URL and ME_API_ROOT_URL are replaced with ME_API_ROOT_URL from the constants.ts
   file. The div element is given the flex and flex-col classes.
 5 In the generateOpenApiSchema.ts file, the USERS_API_FUNCTION_PATH and USERS_API_COLLECTION_FUNCTION_PATH are modified to include the :userName
   parameter.

Overall, these changes seem to be related to updating the API endpoints and making some UI improvements in the codebase.